+++ database/migrations/2019_01_22_012638_create_rewards_table.php
@@ -14,10 +14,10 @@ class CreateRewardsTable extends Migration
     public function up()
     {
         Schema::create('rewards', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('customer_id')->unsigned();
-            $table->integer('job_id')->unsigned()->nullable();
-            $table->integer('payment_id')->unsigned()->nullable();
+            $table->bigIncrements('id');
+            $table->bigInteger('customer_id')->unsigned();
+            $table->bigInteger('job_id')->unsigned()->nullable();
+            $table->bigInteger('payment_id')->unsigned()->nullable();
             $table->integer('value');
             $table->string('description')->nullable();
             $table->timestamps();
